Hi all,
Thank for your suggestions to improve the EHCI emulation patch.  As
this is a new feature it isn't really possible to break it into
smaller patches, but I have as suggested diffed and tested against
trunk, updated the coding style and dropped outlook as a mailer.  I
just tested with a host usb msd device under FC8 and it works fine
with a WinXP target.

This patch does include the mod to usb-linux.c to return STALL instead
of NAK for EPIPE errors as discussed earlier.

Any other comments, suggestions and corrections welcome.

Mark

Rediffed for latest KVM.

Rene Rebe <rene@exactcode.de>

Rediffed for TouchBook QEMU


Signed-off-by: Jason S. McMullan <jason.mcmullan@gmail.com>